extract InstrumentationMapper
