+++ sintr_analysis_server_example/lib/query/completion_metrics.dart
@@ -4,10 +4,7 @@
 
 library sintr_worker_lib.completion;
 
-import 'dart:async';
-
-import 'package:sintr_worker_lib/query.dart';
-import 'package:sintr_worker_lib/session_info.dart';
+import 'package:sintr_worker_lib/instrumentation_query.dart';
 
 const AVE = 'ave';
 const INCOMPLETE = 'incomplete';
@@ -87,9 +84,6 @@ final completionReductionMerge = (Map results1, Map results2) {
   return newResults;
 };
 
-final OPEN_BRACE = '{'.codeUnitAt(0);
-final QUOTE = '"'.codeUnitAt(0);
-
 int _max(int value1, int value2) {
   if (value1 == null) return value2;
   if (value2 == null) return value1;
@@ -138,7 +132,7 @@ void _updateCalculations(sdkResults) {
 
 /// [CompletionMapper] processes session log messages and extracts
 /// metrics for each call to code completion
-class CompletionMapper extends Mapper {
+class CompletionMapper extends InstrumentationMapper {
   /// A mapping of completion notification ID to information abou the completion.
   /// Elements are added when a completion response is found
   /// and removed when the final notification is found.
@@ -149,86 +143,31 @@ class CompletionMapper extends Mapper {
   /// and removed when a matching reponse is found
   Map<String, _Completion> _requestMap = <String, _Completion>{};
 
-  /// The SDK version
-  String _sdkVersion = 'unset';
-
   /// Reporting any partial completion information
   @override
   void cleanup() {
     for (_Completion completion in _requestMap.values) {
       // TODO (danrubel) provide better message for incomplete requests
-      addResult(_sdkVersion, -1);
+      addResult(sdkVersion, -1);
     }
     _requestMap.clear();
     for (_Completion completion in _notificationMap.values) {
       // TODO (danrubel) provide better message for missing notifications
-      addResult(_sdkVersion, -2);
+      addResult(sdkVersion, -2);
     }
     _notificationMap.clear();
   }
 
-  /// Initialize the completion extraction process
   @override
-  Future init(Map<String, dynamic> sessionInfo, AddResult addResult) async {
-    super.init(sessionInfo, addResult);
-    _sdkVersion = sessionInfo[SDK_VERSION] ?? 'unknown';
-  }
-
-  /// Process a log entry and return a string representing the result
-  /// or `null` if no result from the given log entry.
-  @override
-  void map(String logEntryText) {
-    if (logEntryText == null || logEntryText == "") return null;
-    if (!logEntryText.startsWith("~")) return null;
-
-    int index1 = logEntryText.indexOf(':', 1);
-    if (index1 == -1) return null;
-    int time = int.parse(logEntryText.substring(1, index1));
-
-    int index2 = logEntryText.indexOf(':', index1 + 1);
-    if (index2 == -1) return null;
-    String msgType = logEntryText.substring(index1 + 1, index2);
-
-    String message = logEntryText.substring(index2 + 1);
-    _processLogMessage(time, msgType, message);
-  }
-
-  /// Search the given [logMessageText] for the given [key]
-  /// and return the associated value.
-  String _extractJsonValue(String logMessageText, String key) {
-    var prefix = '"$key"::';
-    int start = logMessageText.indexOf(prefix);
-    if (start == -1) throw 'expected $key in $logMessageText';
-    start += prefix.length;
-    int deliminator = logMessageText.codeUnitAt(start);
-    if (deliminator == QUOTE) {
-      // Return quoted string
-      ++start;
-      int end = logMessageText.indexOf('"', start);
-      if (end == -1) throw 'expected value for $key in $logMessageText';
-      return logMessageText.substring(start, end);
-    } else if (deliminator == OPEN_BRACE) {
-      // Return JSON map
-      // TODO (danrubel) handle nested and embedded braces
-      int end = logMessageText.indexOf('}');
-      if (end ==
-          -1) throw 'expected matching brace for $key in $logMessageText';
-      return logMessageText.substring(start, end + 1);
-    }
-    return null;
-  }
-
-  /// Process a log message and return a string representing the result
-  /// or `null` if no result from the given log entry.
-  void _processLogMessage(int time, String msgType, String logMessageText) {
+  void mapLogMessage(int time, String msgType, String logMessageText) {
     if (msgType == 'Req') {
-      String method = _extractJsonValue(logMessageText, 'method');
+      String method = extractJsonValue(logMessageText, 'method');
       _processRequest(time, method, logMessageText);
     } else if (msgType == 'Res') {
-      String requestId = _extractJsonValue(logMessageText, 'id');
+      String requestId = extractJsonValue(logMessageText, 'id');
       _processResponse(time, requestId, logMessageText);
     } else if (msgType == 'Noti') {
-      String event = _extractJsonValue(logMessageText, 'event');
+      String event = extractJsonValue(logMessageText, 'event');
       _processNotification(time, event, logMessageText);
     } else if (msgType == 'Read') {
       // process read entry
@@ -260,7 +199,7 @@ class CompletionMapper extends Mapper {
       if (completion == null) {
         throw 'expected completion request for $logMessageText';
       }
-      addResult(_sdkVersion, time - completion.requestTime);
+      addResult(sdkVersion, time - completion.requestTime);
     }
     return null;
   }
@@ -269,7 +208,7 @@ class CompletionMapper extends Mapper {
   void _processRequest(int time, String method, String logMessageText) {
     // Look for completion requests
     if (method == 'completion.getSuggestions') {
-      String requestId = _extractJsonValue(logMessageText, 'id');
+      String requestId = extractJsonValue(logMessageText, 'id');
       _requestMap[requestId] = new _Completion(time);
     }
   }
@@ -278,8 +217,8 @@ class CompletionMapper extends Mapper {
   void _processResponse(int time, String requestId, String logMessageText) {
     _Completion completion = _requestMap.remove(requestId);
     if (completion != null) {
-      String result = _extractJsonValue(logMessageText, 'result');
-      String notificationId = _extractJsonValue(result, 'id');
+      String result = extractJsonValue(logMessageText, 'result');
+      String notificationId = extractJsonValue(result, 'id');
       _notificationMap[notificationId] = completion;
       completion.responseTime = time;
     }
